Support CLIENT_SECRET_POST as well as POST.
The POST client authentication method is now deprecated and CLIENT_SECRET_POST should be used instead. This doesn’t really change the behaviour, it just allows the new constant to work in the same way as the old one.

We do also throw an error is the configuration is wrong to make it easier to debug when this is the case.